refactor rozenite config to useRozenite hook
